refactor: (#324) CategorySection 컴포넌트에서 카테고리 목록을 불러와 보여주도록 구현
